Massive simplification of filter value display
Settings for the display of filter values were greatly simplified, especially
for number and date property filters. Before, setting how filter values would
be displayed required special properties in the filter pages - to set number
ranges, year vs. month aggregation, etc. Now the code does this automatically,
based on the current layout of data.

This change also allowed for a substantial reduction in the amount of
translation messages, for both special properties and on-screen text.

Change-Id: I6b7cfd37fe96da7289d71a04330f2669d31671a2