UI: show day filters like "February 0, 2019" as month filters instead
This affects the display of "available filters" on Special:BrowseData
when some pages have complete date (year+month+day) and others have
incomplete date (only year+month or even only year).

For pages with incomplete date, the day filter like "February 0, 1555"
will be replaced with month filter "February, 1555"
(or, if month is unknown too, with year filter "1555").

Filters of pages with complete date are unaffected (will continue being
the day filter, as it is now).

Change-Id: I50e0888a39e580bc1146e738f419248d9d02a664