UX: Click on semanticDrilldownCombobox should start Search immediately
On Special:BrowseData, a click on one of the rows in combobox selector
means that the user has chosen the value for this filter.

It should be treated same as the clicks on non-combobox filter options:
immediately open BrowseData with applied filter.
There is no point waiting for user to click Select.

Change-Id: Ie6b2d578bbce80ecb3e4ff0e69b0ad533f900ab4