Fix support for upper/lower date filters in SD_AppliedFilter::checkSQL()
This fixes the filters from SD_BrowseDate::printDateRangeInput(),
which were previously broken and commented out.

(Note: we are not uncommenting printDateRangeInput() for now,
it's lacking in terms of UX and will be improved in a separate commit)

Example of filters that didn't work before this change,
but work with it:

_lower_Date_of_discovery[year]=1900&_upper_Date_of_discovery[day]=1&...
- these querystring parameters are now filtering things correctly.

The problem was that DATE() was called on "o_serialized" (incorrect)
instead of "SUBSTR(o_serialized, 3, 100)" (correct).

Additionally, support for incomplete dates via STR_TO_DATE()
was moved from SDUtils::getDateFunctions(), which only applied to day,
month and year selections,
to earlier SD_Filter::loadDBStructureInformation(),
which is also used for _lower_ and _upper_ filters.

Change-Id: I5a724a2a11142da73e71d99fc1f6d7b7ddcc4b81