Replace $smwgContLang by smwfContLang()
More details. $smwgContLang is not supported by SMW since its release
3.2.0. Instead of $smwgContLang it must be
used smwfContLang(), if not any filter added in a category with the
parser function {{#drilldowninfo}} produces an error when the category
is analized with Special:BrowseData. This commit also introduces a new
public static function (SDUtils::getSMWContLang) to keep the backward
compatibility with previous versions of SMW.

Bug: T265059
Change-Id: Ie1fb46cf7a72ae9c71d360e0bec055a51a0dfdb7