ESLint: remove redundant disable of vars-on-top rule
No longer needed now that we use eslint-config-wikimedia
version 0.20.0, which no longer includes that rule.

If the rule was set to a warning instead of disabled entirely,
I still count that as redundant because the purpose of that
is generally to make the tests not fail (by default they only
fail if a rule triggers an error) rather than to intentionally
opt-in to the rule as a warning.

Bug: T287796
Change-Id: I405d0d3e66e08a804ed477f5435cf868d7d3df3a